tests/wpt/tests/fenced-frame/web-share.https.html

<!DOCTYPE html>
<title>Test of Web Share</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/utils.js"></script>
<script src="/common/utils.js"></script>

<body>

<script>
promise_test(async t => {
  const navigator_share_key = token();

  attachFencedFrame(generateURL('resources/web-share-inner.html',
      [navigator_share_key]));
  const result = await nextValueFromServer(navigator_share_key);

  assert_equals(
      result, 'Web Share failed',
      'Web Share in a fenced frame must fail.');
}, 'Web Sharelock in a fenced frame must fail');
</script>

</body>
